==> 04_Pathway.Rmd <==
---
title: "Reproducing Pathway Figures from the paper"
author: "Sonali Arora"
date: "Sept 23, 2024"
output:
  html_document:
    toc: true
    theme: united
---
In this file, we reproduce all the pathway plots from the paper. 

```{r setup}

rm(list=ls())

library(ggplot2)
library(dplyr)
library(ggpubr)


s1 = 2# size for points in PCA plot
legend_pt_size =4
plot_title_size = 25
axis_text_size = 25
axis_title_size=25
legend_text_size=25
spacing=1
chosen_margin = c(0.5,1,0.5,1)# margins:top,right,bottom,left
theme_clean <- theme_bw() +
    theme(
        plot.title = element_text(hjust=0, vjust=0, 
                                  lineheight=.8, face="bold", size=plot_title_size ),
        plot.margin=unit(chosen_margin,"cm"), 
        legend.text=element_text(size=legend_text_size),
        legend.key.height = unit(spacing, "cm"),
        legend.position = "right",
        legend.justification = 'left',
        legend.spacing.y = unit(0.5, 'cm'),
        legend.title=element_blank() , 
        axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank(), 
        axis.title.y=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank())

myPalette <- colorRampPalette(rev(brewer.pal(11, "Spectral")))
pathwy_sc <- scale_colour_gradientn(colours = myPalette(20), limits=c(-1, 1))

print_mypathway = function(want, want_title){
    midx = sapply(want, function(x)  grep(x, rownames(score))[1] )
    midx = unlist(midx)
    
    L1 = lapply(1:length(midx), function(idx){
        z = midx[idx]
        PATHName = want_title[idx]
        mat = as.data.frame(cbind( umap_data2, GSVA = unlist(score[z,])))      
        
        p1= ggplot(mat, aes(UMAP1_2d, UMAP2_2d, color=GSVA )) +
            geom_point(size=2) + 
            pathwy_sc +
            ggtitle(paste0(PATHName)) + theme_clean + 
            theme(legend.position = "inside", legend.position.inside = c(0.8, 0.3) )
        p1
    })
}
## loading the data 
maindir = "C:/Users/sarora.FHCRC/Desktop/data_for_paper"
resdir = file.path(maindir, "figures_9_19_2024" )

score = readRDS(file.path(maindir, "GSVA_score_v2_only_fetal.rds"))
go_bp = readRDS(file.path(maindir,"GSVA_go_bp_v2.rds"))
identical(colnames(score), colnames(go_bp))
score = rbind(score, go_bp)

umap_data2 = read.delim("umap_data_vst_tpm_umap_9_16_2024.txt", 
                        header=T, stringsAsFactors = FALSE)
score = score[, match(umap_data2$sampleName, colnames(score))]
identical(colnames(score), umap_data2$sampleName) # should be TRUE


umap_data2$Taylor_subgroups = factor(umap_data2$Taylor_subgroups, 
     levels = c("SHH alpha", "SHH beta", "SHH gamma", "SHH delta", 
                "Group3 alpha", "Group3 beta", "Group3 gamma", 
                "Group4 alpha", "Group4 beta", "Group4 gamma", "Not available"))
TAYLOR_cols = c( "darkgoldenrod", "darkgoldenrod1", "darkslateblue", "darkslategray4",  
                 "red", "darkred", "coral", 
                 "seagreen1", "seagreen4", "purple", "grey90")
umap_data2$Taylor_subgroups[which(is.na(umap_data2$Taylor_subgroups))] = "Not available"      
taylor_subtype_plot = ggplot(umap_data2,aes(UMAP1_2d, UMAP2_2d, color=Taylor_subgroups )) +
    geom_point(size=s1) + scale_color_manual(values = TAYLOR_cols, na.value = "grey90") +
    theme_clean +  #ggtitle(paste0("Taylor Medullo classification"))  +
    guides(color=guide_legend(ncol=1))

umap_data2$chosen[which(umap_data2$chosen %in% c("fetal", "WNT", "E1", "E2"))] = NA_character_ 
holland_subtype_plot = ggplot(umap_data2 %>% arrange(desc(is.na(chosen))),
                              aes(UMAP1_2d, UMAP2_2d, color=chosen )) +
    geom_point(size=s1) + scale_color_manual(values = col_vector[1:11], na.value = "grey90") +
    theme_clean +  #ggtitle(paste0("Taylor Medullo classification"))  +
    guides(color=guide_legend(ncol=1))

## shh main figure
chosen_pathways = c("GLI_PROTEINS_BIND_PROMOTERS", 
                    "RUNX3_REGULATES_YAP1_MEDIATED", 
                    "KEGG_RIBOSOME", 
                    "BIOCARTA_BLYMPHOCYTE_PATHWAY", 
                    "REGULATION_OF_PROTEIN_KINASE_C_ACTIVITY", 
                    "REGULATION_OF_T_CELL_ACTIVATION_VIA_T_CELL")

wtitle_name = c("Gli  binds HH promoters", "RUNX3 regulated YAP1", 
                "Ribosome", "Blymphocyte pathway", "Protein Kinase C Activity", 
                "T cell activation")
k1 = print_mypathway(want  = chosen_pathways, want_title = wtitle_name )
final_plot <- ggarrange(plotlist = c(list(holland_subtype_plot, 
    taylor_subtype_plot + guides(colour = guide_legend(override.aes = list(size=4))), NULL),
                                     k1), ncol = 3 , nrow =3)

pdf(file.path(resdir, "Multiplot_fig3_9_16_2024_9panel.pdf"),width =23, height = 16)
print(final_plot)
dev.off()

## shh supp fig
chosen_pathways = c("GO_POSITIVE_REGULATION_OF_VASCULAR_WOUND_HEA", 
                    "GO_REGULATION_OF_COMPLEMENT_ACTIVATION", 
                    "GO_HOMOSERINE_METABOLIC_PROCESS", 
                    "REACTOME_RUNX2_REGULATES_CHONDROCYTE_MATURATION",
                    "REACTOME_RUNX1_REGULATES_TRANSCRIPTION_OF_GENES_INVOLVED_IN_INTERLEUKIN_SIGNALING",
                    "REACTOME_NONSENSE_MEDIATED_DECAY_NMD")
wtitle_name = c("WOUND HEALING", "Complement activation", 
                "Metabolic process", "Runx2", "Runx1", 
                "NMD")
k2 = print_mypathway(want  = chosen_pathways, want_title = wtitle_name )
SUPP_plot <- ggarrange(plotlist = c(list(holland_subtype_plot, 
     taylor_subtype_plot + guides(colour = guide_legend(override.aes = list(size=4))), NULL),
                                     k2), ncol = 3 , nrow =3)

pdf(file.path(resdir, "Multiplot_SUPP_fig3_9_16_2024_9panel.pdf"),width =23, height = 16)
print(SUPP_plot)
dev.off()

## g3 / g4  main fig 4
chosen_pathways = c("REACTOME_EUKARYOTIC_TRANSLATION_ELONGATION", 
                    "REACTOME_ACTIVATION_OF_THE_PHOTOTRANSDUCTION", 
                    "REACTOME_ACTIVATED_NTRK2_SIGNALS_THROUGH_CDK", 
                    "REACTOME_PRESYNAPTIC_DEPOLARIZATION_AND_CALCIUM",
                    "REACTOME_STAT5_ACTIVATION",
                    "REACTOME_SIGNALING_BY_KIT_IN_DISEASE")

g1 = print_mypathway(want  = chosen_pathways, want_title = chosen_pathways )
main4 <- ggarrange(plotlist = g1, ncol = 2 , nrow =3)
pdf(file.path(resdir, "Multiplot_main_fig4_9_16_2024_9panel.pdf"),width =14, height = 18)
print(main4)
dev.off()

## g3/g4 supp fig
chosen_pathways = c(
                    "BIOCARTA_TERT_PATHWAY", 
                    "BIOCARTA_WNT_LRP6_PATHWAY",                   
                    "GO_REGULATION_OF_SMAD_PROTEIN_COMPLEX_ASSEMBLY",
                    "GO_REGULATION_OF_VOLTAGE_GATED_POTASSIUM", 
                    "GO_REGULATION_OF_INTERLEUKIN_6_MEDIATED_SIGNALING_PATHWAY",                    
                    "REACTOME_SIGNALING_BY_LEPTIN", 
                    "BIOCARTA_IL22BP_PATHWAY", 
                    "REACTOME_NTRK2_ACTIVATES_RAC1",
                    "GO_HISTONE_H3_K27_TRIMETHYLATION")

g2 = print_mypathway(want  = chosen_pathways, want_title = chosen_pathways )
supp4 <- ggarrange(plotlist = g2, ncol = 3 , nrow =3)
pdf(file.path(resdir, "Multiplot_SUPP_fig4_9_16_2024_9panel.pdf"),width =21, height = 18)
print(supp4)
dev.off()

## main fig 6  - e1 and e2 pathways
chosen_pathways = c(
    "BIOCARTA_P53_PATHWAY", 
    "ACTIVATION_OF_RAS_IN_B_CELLS",
    "G1_S_SPECIFIC_TRANSCRIPTION",
    "CHROMATIN_MAINTENANCE", 
    "DOPAMINE_RECEPTORS",
    "BIOCARTA_RHODOPSIN_PATHWAY",
    "REGULATION_OF_CARDIAC_VASCULAR_SMOOTH_MUSCLE_CELL", 
    "ANTIGEN_PROCESSING_AND_PRESENTATION_VIA_MHC_CLASS")

m1 = print_mypathway(want  = chosen_pathways, want_title = chosen_pathways )
main6 <- ggarrange(plotlist = m1, ncol = 2 , nrow =4)
pdf(file.path(resdir, "Multiplot_main_fig6_9_16_2024_9panel.pdf"),width =14, height = 21)
print(main6)
dev.off()
